requirements.txt
pyyaml
python-gitlab
pybitbucket
pygithub3
multiprocess